Query: Remove EntityQueryableExpressionVisitor and integrate functionality in QueryableMethodTranslatingVisitor
Issue:
We did 2 pass in ExpressionTree to convert entityQueryables to shapedQueries and then translate/compose over them.
Since we mutate the ShapedQueryExpression.QueryExpression in place,
(reference to query expression are in shaper expression so it is not easy to make it immutable)
if we try to translate something and fail then there is no way for us to try translating smaller part since we already mutated the ET.
By integrating functionality in translation pipeline, if we fail to translate, our initial expression would still be intact (since it does not contain shapedQuery

This is required for supporting collections/single non-scalar in projection.
