+++ src/EFCore.Relational/Query/Pipeline/RelationalQueryableMethodTranslatingExpressionVisitor.cs
@@ -46,6 +46,19 @@ namespace Microsoft.EntityFrameworkCore.Relational.Query.Pipeline
             _sqlExpressionFactory = sqlExpressionFactory;
         }
 
+        protected override Expression VisitMethodCall(MethodCallExpression methodCallExpression)
+        {
+            if (methodCallExpression.Method.DeclaringType == typeof(RelationalQueryableExtensions)
+                && methodCallExpression.Method.Name == nameof(RelationalQueryableExtensions.FromSqlOnQueryable))
+            {
+                var sql = (string)((ConstantExpression)methodCallExpression.Arguments[1]).Value;
+                var queryable = (IQueryable)((ConstantExpression)methodCallExpression.Arguments[0]).Value;
+                return CreateShapedQueryExpression(queryable.ElementType, sql, methodCallExpression.Arguments[2]);
+            }
+
+            return base.VisitMethodCall(methodCallExpression);
+        }
+
         public override ShapedQueryExpression TranslateSubquery(Expression expression)
         {
             return (ShapedQueryExpression)new RelationalQueryableMethodTranslatingExpressionVisitor(
@@ -53,6 +66,35 @@ namespace Microsoft.EntityFrameworkCore.Relational.Query.Pipeline
                 _sqlTranslator,
                 _sqlExpressionFactory).Visit(expression);
 
+        }
+
+        protected override ShapedQueryExpression CreateShapedQueryExpression(Type elementType)
+        {
+            var entityType = _model.FindEntityType(elementType);
+            var queryExpression = _sqlExpressionFactory.Select(entityType);
+
+            return CreateShapedQueryExpression(entityType, queryExpression);
+        }
+
+        private ShapedQueryExpression CreateShapedQueryExpression(Type elementType, string sql, Expression arguments)
+        {
+            var entityType = _model.FindEntityType(elementType);
+            var queryExpression = _sqlExpressionFactory.Select(entityType, sql, arguments);
+
+            return CreateShapedQueryExpression(entityType, queryExpression);
+        }
+
+        private ShapedQueryExpression CreateShapedQueryExpression(IEntityType entityType, SelectExpression selectExpression)
+        {
+            return new ShapedQueryExpression(
+                selectExpression,
+                new EntityShaperExpression(
+                entityType,
+                new ProjectionBindingExpression(
+                    selectExpression,
+                    new ProjectionMember(),
+                    typeof(ValueBuffer)),
+                false));
         }
 
         protected override ShapedQueryExpression TranslateAll(ShapedQueryExpression source, LambdaExpression predicate)
